# docs/swagger.yaml
basePath: /api/v1
definitions:
  checklist.CheckResult:
    properties:
      description:
        type: string
      details:
        type: string
      result:
        type: string
      timestamp:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server for OpenStack security checking
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenStack Security Hub API
  version: "1.0"
paths:
  /check/identity-01:
    get:
      description: Execute all Identity-01 series security checks for OpenStack
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/checklist.CheckResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Run all Identity-01 security checks
      tags:
      - identity
  /check/identity-01-01:
    get:
      description: Check Identity configuration security
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/checklist.CheckResult'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Run Identity-01-01 check
      tags:
      - identity
  /check/identity-01-02:
    get:
      description: Check Identity configuration security
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/checklist.CheckResult'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Run Identity-01-02 check
      tags:
      - identity
swagger: "2.0"